forced correct type for winnowing density test

+++ tests/test_winnowing.py
@@ -59,7 +59,8 @@ class WinnowDensityTestCase(unittest.TestCase):
         for win_size in range(2, 20):
             expected_density = 2/(win_size + 1)*10000
             rand_hashes = np.random.randint(-9223372036854775808,
-                                             9223372036854775807, 10000)
+                                             9223372036854775807, 10000,
+                                             dtype=np.int64)
             density = len(_winnow(rand_hashes, win_size))
 
             if not (density > expected_density - 200 and
